update validations conditions
added conditions for validations. we can now create an empty BlogPost if
published is not true. validations is required only if published is true

added a random title on creation so before_save doesn't fail

added methods for safe_body and safe_description that returns an empty
string if they're nil or return the text escaped

added specs and updated the factory