created the pictures controller
created the routes for the pictures controller. They will have the
model_name, model_id and for destroy the id of the picture that will be
destroyed.

overwritten the index, create and destroy methods. added changes to take
into account the polymorphic join table. for now we only have the
BlogPost. when the newsletter will be present we can send that to the
controller and it should work